se separa la clase venta en un archivo aparte actualización del README a Practica 5
